+++ src/hooks/useDetectClose.tsx
